Adds account support to financial tracking system
Introduces comprehensive account management functionality to enable users to track financial transactions across different accounts like checking, savings, and credit cards.

Creates new Account model with support for various account types, provider information, and last four digits for privacy. Updates transaction model to optionally link to accounts and enhances CSV import to automatically create accounts from imported data.

Adds dedicated accounts API endpoints for creating and retrieving accounts, along with corresponding React hooks and UI components for account management integration throughout the expense tracking interface.

Improves transaction display by showing associated account information and refactors UI components for better reusability across the application.

+++ src/components/dashboard/CurrentFinaces.tsx
@@ -12,10 +12,9 @@ function CurrentFinances() {
 
   return (
     <span>
-      {transactions.reduce(
-        (acc, transaction) => acc + parseInt(transaction.amount),
-        0
-      ).toLocaleString()}
+      {transactions
+        .reduce((acc, transaction) => acc + parseFloat(transaction.amount), 0)
+        .toLocaleString()}
     </span>
   );
 }
